Refactoring parsing, rendering, text editor and interactors

+++ src/building.hpp
@@ -8,23 +8,23 @@
 class Building {
 protected:
 public:
-    Building(int x, int y, int w, int h, int id, tcod::ColorRGB color, Map& map):
+    Building(long x, long y, int w, int h, int id, tcod::ColorRGB color, Map& map):
         x(x), y(y), w(w), h(h), id(id), color(color)
     {
         // Mark building location in map
-        for( int xi = x; xi < x + w; ++xi ) {
-            for( int yi = y; yi < y + h; ++yi ) {
+        for( long xi = x; xi < x + w; ++xi ) {
+            for( long yi = y; yi < y + h; ++yi ) {
                 map.getTile(xi, yi).bldgId.emplace(id);
             }
         }
     }
-    const int x, y, w, h, id; // bottom left (x, y), width, height
+    const long x, y, w, h, id; // bottom left (x, y), width, height
     const tcod::ColorRGB color;
 };
 
 class Nursery : public Building {
 public:
-    Nursery(int x, int y, int id, Map& map): Building(x, y, 3, 3, id, color::blue, map) {}
+    Nursery(long x, long y, int id, Map& map): Building(x, y, 3, 3, id, color::blue, map) {}
 
 };
